+++ spine-corona/main.lua
@@ -5,6 +5,8 @@ local spine = require "spine-corona.spine"
 local skeletons = {}
 local activeSkeleton = 1
 local lastTime = 0
+local swirl = spine.SwirlEffect.new(400)
+local swirlTime = 0
 
 function loadSkeleton(atlasFile, jsonFile, x, y, scale, animation, skin)
 	-- to load an atlas, we need to define a function that returns
@@ -67,7 +69,11 @@ function loadSkeleton(atlasFile, jsonFile, x, y, scale, animation, skin)
 		animationState:setAnimationByName(0, "walk", true)
 		local jumpEntry = animationState:addAnimationByName(0, "jump", false, 3)
 		animationState:addAnimationByName(0, "run", true, 0)
-  else
+  elseif atlasFile == "raptor.atlas" then
+		--skeleton.vertexEffect = spine.JitterEffect.new(5, 5)
+		skeleton.vertexEffect = swirl
+		animationState:setAnimationByName(0, animation, true)
+	else
     animationState:setAnimationByName(0, animation, true)
   end
 
@@ -77,10 +83,10 @@ end
 
 -- table.insert(skeletons, loadSkeleton("coin.atlas", "coin-pro.json", 240, 300, 0.4, "rotate"))
 -- table.insert(skeletons, loadSkeleton("spineboy.atlas", "spineboy-ess.json", 240, 300, 0.4, "walk"))
--- table.insert(skeletons, loadSkeleton("raptor.atlas", "raptor-pro.json", 200, 300, 0.25, "walk"))
+table.insert(skeletons, loadSkeleton("raptor.atlas", "raptor-pro.json", 200, 300, 0.25, "walk"))
 -- table.insert(skeletons, loadSkeleton("goblins.atlas", "goblins-pro.json", 240, 300, 0.8, "walk", "goblin"))
 -- table.insert(skeletons, loadSkeleton("stretchyman.atlas", "stretchyman-pro.json", 40, 300, 0.5, "sneak"))
-table.insert(skeletons, loadSkeleton("tank.atlas", "tank-pro.json", 400, 300, 0.2, "drive"))
+-- table.insert(skeletons, loadSkeleton("tank.atlas", "tank-pro.json", 400, 300, 0.2, "drive"))
 -- table.insert(skeletons, loadSkeleton("vine.atlas", "vine-pro.json", 240, 300, 0.3, "grow"))
 
 local triangulator = spine.Triangulator.new()
@@ -109,6 +115,11 @@ Runtime:addEventListener("enterFrame", function (event)
 	local currentTime = event.time / 1000
 	local delta = currentTime - lastTime
 	lastTime = currentTime
+	
+	swirlTime = swirlTime + delta
+	local percent = swirlTime % 2
+	if (percent > 1) then percent = 1 - (percent - 1) end
+	swirl.angle = spine.Interpolation.apply(spine.Interpolation.pow2, -60, 60, percent)
 
 	skeleton = skeletons[activeSkeleton].skeleton
 	skeleton.group.isVisible = true